requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic[email]>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
aiosqlite>=0.20.0
greenlet>=3.0.0
alembic>=1.11.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
